Move supportsInclude to internal API
